Register Page for iOS
I figured we needed a register page in order to setup the login. I attached a register page storyboard created in swift so this is what the actual view would look like in app. I wasn't sure how to implement the various methods of registration (Github, email, Slack, etc.), so I left it at email. There's the possibility of using a "picker" object which acts as a type of dropdown menu where the user could choose the platform they wish to register with.